--- a/client/client.go
+++ b/client/client.go
@@ -92,8 +92,8 @@ func newRPC2Client(conn net.Conn, database *model.DBModel) (*OvsdbClient, error)
 	ovs.rpcClient.Handle("echo", func(_ *rpc2.Client, args []interface{}, reply *[]interface{}) error {
 		return ovs.echo(args, reply)
 	})
-	ovs.rpcClient.Handle("update", func(_ *rpc2.Client, args []interface{}, _ *[]interface{}) error {
-		return ovs.update(args)
+	ovs.rpcClient.Handle("update", func(_ *rpc2.Client, args []interface{}, reply *[]interface{}) error {
+		return ovs.update(args, reply)
 	})
 	go ovs.rpcClient.Run()
 	go ovs.handleDisconnectNotification()
@@ -189,12 +189,11 @@ func (ovs *OvsdbClient) echo(args []interface{}, reply *[]interface{}) error {
 
 // RFC 7047 : Update Notification Section 4.1.6
 // Processing "params": [<json-value>, <table-updates>]
-func (ovs *OvsdbClient) update(params []interface{}) error {
+func (ovs *OvsdbClient) update(params []interface{}, reply *[]interface{}) error {
 	if len(params) < 2 {
 		return fmt.Errorf("invalid update message")
 	}
 	// Ignore params[0] as we dont use the <json-value> currently for comparison
-
 	raw, ok := params[1].(map[string]interface{})
 	if !ok {
 		return fmt.Errorf("invalid update message")
@@ -216,7 +215,7 @@ func (ovs *OvsdbClient) update(params []interface{}) error {
 	for _, handler := range ovs.handlers {
 		handler.Update(params[0], tableUpdates)
 	}
-
+	*reply = []interface{}{}
 	return nil
 }