--- a/mapper/mapper.go
+++ b/mapper/mapper.go
@@ -286,7 +286,7 @@ func (m Mapper) NewCondition(tableName string, data interface{}, field interface
 
 // newMutation creates a RFC7047 mutation object based on an ORM object and the mutation fields (in native format)
 // It takes care of field validation against the column type
-func (m Mapper) NewMutation(tableName string, data interface{}, column string, mutator ovsdb.Mutator, value interface{}) ([]interface{}, error) {
+func (m Mapper) NewMutation(tableName string, data interface{}, column string, mutator ovsdb.Mutator, value interface{}) (*ovsdb.Mutation, error) {
 	table := m.Schema.Table(tableName)
 	if table == nil {
 		return nil, newErrNoTable(tableName)
@@ -325,7 +325,7 @@ func (m Mapper) NewMutation(tableName string, data interface{}, column string, m
 		}
 	}
 
-	return []interface{}{column, mutator, ovsValue}, nil
+	return &ovsdb.Mutation{Column: column, Mutator: mutator, Value: ovsValue}, nil
 }
 
 // equalIndexes returns whether both models are equal from the DB point of view