This PR builds on #122 

1. Simplifies the marshal/unmarshal of the structs in the `update` code path (TableUpdates, TableUpdate)
1. Adds some missing fields in the `Operation` struct
1. Adds an `Echo` method on the `client` for calling the server
1. Ensures that our `update` method response contains either an error or a result to avoid upsetting some RPC clients.
1. Uses partial unmarshalling via `json.RawMessage` in the `update` arguments so we can avoid having to marshal/unmarshal to get the arguments in to the right type
1. Re-writes the benchmarks to use `[]byte`


Comparison of benchmarks:
```
name       old time/op  new time/op  delta
Update1-8  43.0µs ± 1%  14.6µs ± 1%  -65.96%  (p=0.000 n=3+3)
Update2-8  59.2µs ± 3%  21.6µs ± 2%  -63.59%  (p=0.000 n=3+3)
Update3-8  77.3µs ± 1%  28.6µs ± 1%  -63.02%  (p=0.000 n=3+3)
Update5-8   108µs ± 2%    43µs ± 2%  -60.71%  (p=0.000 n=3+3)
Update8-8   157µs ± 2%    64µs ± 1%  -59.29%  (p=0.000 n=3+3)
```

In terms of real world performance, with `stress` we're ~6 seconds better on 3000 inserts

**Before**
```
$ go run ./cmd/stress -verbose -ninserts 3000 -ovsdb tcp:localhost:49154
rootUUID is 2333c4e6-5de6-4c8f-a7ba-3f62557d3957

Summary:
        Insertions: 3000
        Deletions: 0
go run ./cmd/stress -verbose -ninserts 3000 -ovsdb tcp:localhost:49154  87.58s user 2.94s system 67% cpu 2:14.57 total
```

*After*
```
$ time go run ./cmd/stress -verbose -ninserts 3000 -ovsdb tcp:localhost:49153 
rootUUID is 28ddc871-f67a-405c-ac47-7beb6b94526a
Summary:
        Insertions: 3000
        Deletions: 0
go run ./cmd/stress -verbose -ninserts 3000 -ovsdb tcp:localhost:49153  78.19s user 2.54s system 63% cpu 2:08.10 total
```

Fixes: #109