@dave-tucker 2021-05-25
@amorenoz rebased and retested.
without stress on 3000 operations completes in 1m49s, with the changes it's 1m43s which is consistent with the ~6sec improvement I was seeing before.